pyproject.toml
[project]
name = "web-use"
version = "0.1.0"
description = "An AI Agent that automates web browsing tasks."
authors = [
    { name = "Jeomon George", email = "jeogeoalukka@gmail.com" }
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.13"

dependencies = [
    "chromedeveloperprotocol-client @ file:///D:/Projects/CDP-Client",
    "anthropic>=0.68.1",
    "google-genai>=1.45.0",
    "openai>=1.93.0",
    "groq>=0.29.0",
    "ollama>=0.5.1",
    "cerebras-cloud-sdk>=1.50.1",
    "litellm>=1.72.0",
    "mistralai>=1.9.11",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "termcolor>=3.1.0",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.1",
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
packages = ["src"]
